requirements.txt
psycopg2-binary==2.9.3
PyYAML==6.0
requests==2.28.1
redis==4.3.4